Parametrize baseURL for PMWEBD
Supports a usecase where one might not be to use any other ports
except port 80, and might need to set up a reverse-proxy to proxy
the request to the pmwedbd.

Example, set up the web server hosting Vector application with a
reverse-proxy to route HTTP requests starting with the baseURL '/pmweb'
to an internal pmwebd on a different port, which can be used in combination
with the hostspec mechanism to query different machines behind the proxy.